fix: remove duplicated hybrid-scoring and Thai-normalization logic
Two independent implementations of the same business rules had
drifted into separate files, risking silent divergence:

- src/api/routers/classify.py::_classify_hybrid() reimplemented the
  keyword/embedding weighting formula inline (KEYWORD_WEIGHT=0.60,
  EMBEDDING_WEIGHT=0.40) instead of calling the existing
  src/core/scoring_logic.py::calculate_hybrid_score(). Now it does.

- taxonomy-app had two byte-identical copies of a ThaiTextProcessor
  class (process-local/route.ts, process-storage/route.ts) whose
  clean() only stripped characters/whitespace, unlike the backend's
  ThaiTextProcessor (src/core/fresh_implementations.py) which also
  lowercases, fixes floating-vowel ordering (เ็→เ, แ็→แ), and
  converts Thai digits to Arabic digits before cleaning. Extracted
  a single taxonomy-app/utils/thai-text-processor.ts that both
  routes now import, with clean() porting the same
  lowercase -> vowel-fix -> digit-conversion -> strip order so a
  product's cleaned tokens/keywords no longer depend on which
  pipeline (frontend fallback vs backend) touched it.
