+++ taxonomy-app/app/api/import/process-local/route.ts
@@ -1,61 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { createClient } from '@supabase/supabase-js'
+import { ThaiTextProcessor } from '@/utils/thai-text-processor'
 
 const supabase = createClient(
   process.env.SUPABASE_URL || 'http://127.0.0.1:54331',
   process.env.SUPABASE_SERVICE_ROLE_KEY || process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
 )
-
-// Thai text processing utilities (same as before)
-class ThaiTextProcessor {
-  static clean(text: string): string {
-    return text
-      .replace(/[^\u0E00-\u0E7Fa-zA-Z0-9\s\-\.\(\)]/g, '')
-      .replace(/\s+/g, ' ')
-      .trim()
-      .toLowerCase()
-  }
-
-  static tokenize(text: string): string[] {
-    const tokens = text
-      .split(/[\s\-\(\)\[\]\/\\,\.]+/)
-      .filter(token => token.length >= 2)
-    return [...new Set(tokens)]
-  }
-
-  static extractUnits(text: string): string[] {
-    const unitPatterns = [
-      /(\d+)\s*(กรัม|g|gram)/gi,
-      /(\d+)\s*(มิลลิลิตร|ml|มล)/gi,
-      /(\d+)\s*(ลิตร|l)/gi,
-      /(\d+)\s*(กิโลกรม|kg|กก)/gi,
-      /(\d+)\s*(ชิ้น|pcs)/gi,
-      /(\d+)\s*(แพ็ค|pack)/gi,
-      /(\d+)\s*(กล่อง|box)/gi
-    ]
-
-    const units: string[] = []
-    unitPatterns.forEach(pattern => {
-      const matches = text.match(pattern)
-      if (matches) units.push(...matches)
-    })
-    return units
-  }
-
-  static extractAttributes(text: string): Record<string, any> {
-    const attributes: Record<string, any> = {}
-    
-    const colors = ['แดง', 'เขียว', 'น้ำเงิน', 'เหลือง', 'ขาว', 'ดำ', 'ชมพู', 'ม่วง', 'ส้ม', 'เทา']
-    const foundColors = colors.filter(color => text.includes(color))
-    if (foundColors.length > 0) attributes.colors = foundColors
-
-    const sizes = ['S', 'M', 'L', 'XL', 'XXL', 'เล็ก', 'กลาง', 'ใหญ่']
-    const foundSizes = sizes.filter(size => text.includes(size))
-    if (foundSizes.length > 0) attributes.sizes = foundSizes
-
-    return attributes
-  }
-}
 
 /**
  * Generate embedding using local model via Edge Function
